requirements.txt
asgiref==3.6.0
autopep8==2.0.1
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
crispy-bootstrap5==0.7
cssselect2==0.8.0
Django==4.1.7
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-js-asset==2.0.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
fonttools==4.61.1
idna==3.11
phonenumbers==8.13.6
pillow==12.0.0
pycodestyle==2.10.0
pycparser==2.23
pydyf==0.12.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
qrcode==8.2
requests==2.32.5
six==1.17.0
soupsieve==2.8.1
sqlparse==0.4.3
tinycss2==1.5.1
tinyhtml5==2.0.0
tomli==2.0.1
typing_extensions==4.15.0
tzdata==2022.7
urllib3==2.6.2
weasyprint==67.0
webencodings==0.5.1
zopfli==0.4.0
